We can get more precision when we propagate across a pop
When coming down from a pop, we know that the min headroom is no smaller than
the next flow of the pop's flow value. However, we also know that whatever level
we're popping onto has a minimum. Whichever minimum is higher is the correct
one. We know this because we might have lost this extra precision when two calls
had to merge.